threshold: switch to cidr based ip parsing
Additionally minor code cleanups.
