+++ src/detect.h
@@ -715,7 +715,19 @@ typedef struct DetectEngineLookupFlow_ {
     struct SigGroupHead_ *sgh[256];
 } DetectEngineLookupFlow;
 
-#include "detect-threshold.h"
+typedef struct DetectThresholdEntry_ {
+    uint32_t sid; /**< Signature id */
+    uint32_t gid; /**< Signature group id */
+
+    uint32_t tv_timeout;    /**< Timeout for new_action (for rate_filter)
+                                 its not "seconds", that define the time interval */
+    uint32_t seconds;       /**< Event seconds */
+    uint32_t current_count; /**< Var for count control */
+    int track;              /**< Track type: by_src, by_src */
+
+    struct timeval tv1; /**< Var for time control */
+    struct DetectThresholdEntry_ *next;
+} DetectThresholdEntry;
 
 /** \brief threshold ctx */
 typedef struct ThresholdCtx_    {
